updated pyplot demo
