+++ tk_builder/panels/pyplot_panel/pyplot_panel.py
@@ -1,10 +1,7 @@
-from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
-from matplotlib.figure import Figure
 from matplotlib.collections import LineCollection
 import time
 import numpy as np
 
-import tkinter
 from tk_builder.widgets import basic_widgets
 from tk_builder.panels.pyplot_panel.pyplot_panel_utils.plot_style_utils import PlotStyleUtils
 from tk_builder.panel_builder.widget_panel import WidgetPanel
@@ -16,28 +13,23 @@ SCALE_Y_AXIS_PER_FRAME_FALSE = "don't scale y axis per frame"
 PYPLOT_UTILS = PlotStyleUtils()
 
 
-class PyplotCanvas(tkinter.LabelFrame):
-    def __init__(self, primary):
-        tkinter.LabelFrame.__init__(self, primary)
-        fig = Figure()
-        self.ax = fig.add_subplot(111)
-
-        self.canvas = FigureCanvasTkAgg(fig, primary)
-        self.canvas.get_tk_widget().pack(fill='both')
-
-
 class PyplotControlPanel(WidgetPanel):
-    widget_list = ("color_palette_label", "color_palette", "n_colors_label", "n_colors",
-                   "scale",
-                   "rescale_y_axis_per_frame",
-                   "fps_label", "fps_entry", "animate")
+    _widget_list = ("color_palette_label",
+                    "color_palette",
+                    "n_colors_label",
+                    "n_colors",
+                    "scale",
+                    "rescale_y_axis_per_frame",
+                    "fps_label",
+                    "fps_entry",
+                    "animate")
     color_palette_label = widget_descriptors.LabelDescriptor("color_palette_label")  # type: basic_widgets.Label
     color_palette = widget_descriptors.ComboboxDescriptor("color_palette")                     # type: basic_widgets.Combobox
     n_colors_label = widget_descriptors.LabelDescriptor("n_colors_label")                        # type: basic_widgets.Label
     n_colors = widget_descriptors.SpinboxDescriptor("n_colors")                            # type: basic_widgets.Spinbox
     scale = widget_descriptors.ScaleDescriptor("scale")                                 # type: basic_widgets.Scale
-    rescale_y_axis_per_frame = widget_descriptors.ComboboxDescriptor("")          # type: basic_widgets.Combobox
-    animate = widget_descriptors.ButtonDescriptor("rescale_y_axis_per_frame")                             # type: basic_widgets.Button
+    rescale_y_axis_per_frame = widget_descriptors.ComboboxDescriptor("rescale_y_axis_per_frame")          # type: basic_widgets.Combobox
+    animate = widget_descriptors.ButtonDescriptor("animate")                             # type: basic_widgets.Button
     fps_label = widget_descriptors.LabelDescriptor("fps_label")                             # type: basic_widgets.Label
     fps_entry = widget_descriptors.EntryDescriptor("fps_entry")                            # type: basic_widgets.Entry
 
@@ -75,16 +67,16 @@ class AppVariables():
 
 
 class PyplotPanel(WidgetPanel):
-    pyplot_canvas = PyplotCanvas           # type: PyplotCanvas
-    control_panel = PyplotControlPanel      # type: PyplotControlPanel
+    _widget_list = ("pyplot_canvas", "control_panel")
+    pyplot_canvas = widget_descriptors.PyplotCanvasDescriptor("pyplot_canvas")           # type: PyplotCanvas
+    control_panel = widget_descriptors.PanelDescriptor("control_panel", PyplotControlPanel)  # type: PyplotControlPanel
 
     def __init__(self, primary):
         WidgetPanel.__init__(self, primary)
 
         self.variables = AppVariables()
         self.pyplot_utils = PlotStyleUtils()
-        widget_list = ["pyplot_canvas", "control_panel"]
-        self.init_w_vertical_layout(widget_list)
+        self.init_w_vertical_layout()
 
         canvas_size_pixels = self.pyplot_canvas.canvas.figure.get_size_inches() * self.pyplot_canvas.canvas.figure.dpi
